update for jsx folders and little changes
