--- my-app/src/components/badge/Badge.js
@@ -1,11 +1,0 @@
-import "./Badge.css"
-function Badge(props) {
-    return (
-      <div className="badge">
-        <p>{props.status}</p>
-      </div>
-    );
-}
-
-
-export default Badge
