+++ my-app/src/components/task-viewer/TaskViewer.jsx
@@ -1,3 +1,4 @@
+import React from "react";
 import "./TaskViewer.css";
 import TaskCard from "../task-card/TaskCard";
